integrate viewmodel
